Refactoring of the Ship object
* Conversion to the factory function pattern
* Unit tests added

`Ship` is now a factory function, instead of an object with factory abilities. It takes a DOM Element object and augments it with the properties that constitute a ship: a word, a type, etc.

Since it is not an object anymore, the `locate` and `_shipIdFromWord` functions are gone; these responsabilities will have to be re-abstracted away to some other object later (ie. part of the code is less DRY now, but it's for the best).