Refactoring of the Dockyard object
* Conversion to the Factory Function pattern
* Unit tests added

`Dockyard` is now a factory function, instead of an object with factory abilities. It takes a DOM Element object and augments it with the properties that constitute a dockyard: `build()` to build a ship, `launch()` to add it to a combat zone, `locateShip` to find a given ship (through its word), etc.

The HTML templates for the ships (which are used by the `dockyard` objects to build the ships) are not `<template>` HTML5 elements instead of `<div>`s. This is probably is bit nicer (and more semantic), and certainly edgier, but also less compatible with Phantomjs. Since I couldn't find a good poylfill that works with Phantomjs, I've switched the driver to Selenium.