The `Game` object is decorated before being rendered
I'd like to let the rendered object (`@game) to tells us how it wants to
be rendered, instead of letting the controller decide (ie. leaving the
default call to `render` in the controller action, and calling `render @game`
from the view). To do so, we have to override the default method
`to_partial_path` from `ActiveModel::Conversion`.

However, the choice of the relevant partial depends on both the game's
current state (is it running, over, etc.) **and** the current context,
namely, _who_ is looking at the game: is it one of the current players,
a new player waiting to join the game, or a guest.

Since we don't want the `Game` model to be context-aware, we need a new object
to handle this responsibility; something that would represent a game _as seen
by a certain person_. This looks like a good role for a _decorator_.

I chose to implement this decorator with the simplest object possible: a
`SimpleDelegator`. Since we're following Rails' convention of naming objects
(expect for models) after their _nature_ (ie. `foo_controller`, `foo_helper`,
etc.), I named this object `GameExhibit`.

(I chose the `Exhibit` suffix after Avdi Grimm's
[Objects on Rails](http://objectsonrails.com/#ID-2656c30c-080a-4a4e-a53e-4fbaad39c262),
where Avdi makes a good case against using more familiar terms such as `Decorator`
or `Presenter`.)

A note on the specs: I tried to follow the apparent convention of large contexts
and minimal expectations. I'm not entirely convinced (it looks a bit too cryptic
and magical to me), so feedback is especially welcome here.