Refactoring of the GameExhibit
Asking the game object if it is in a given state instead of checking the
actual value of its `:state` attribute allows for a friendlier DSL (and,
more importantly, provides some nice encapsulation). The relevant methods
themselves (eg. `Game#running?`), as well as methods to _set_ the state
(eg. `Game#running!`) are meta-programmed, but I don't think that it's too
edgy or tricky to be safe.

I've also added a bit of responsibility to the `Player` model, namely
being able to tell if it is part of a given game. This is more of a convenience
method than a critical piece of business logic, but hey, it makes for a nice DSL!