Fixed: ships can appear too high or too low on the screen
Ships were added to the `#combat_zone` div, which used to include the actual “combat zone”, the stars-scrolling background, **and** the HUD (which was **not** actually inside the `hud` div).

To reduce the confusion, I've added a div around the HUD elements, and renamed everything:

*   `combat_zone` (actual zone + hud + background) is now `game_window`
*   `hud` (actually the combat zone + the HUD elements) is now `combat_zone`
*   the HUD elements are not in a `hud` div

(I didn't date move the `hud` _outside_ the `combat_zone`, even though it would have made things easier.)

Now, the space occupied by the HUD can more easily be taken into account when computing the available vertical space in the combat zone, when adding a ship to it. To do so, I've simply cannibalised the code from #101.